pyproject.toml
[project]
name = "portfolio-cutdown"
version = "0.1.0"
description = "Personal portfolio CMS"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "beautifulsoup4>=4.14.3",
  "boto3>=1.42.38",
  "fastapi>=0.128.0",
  "jinja2>=3.1.6",
  "markdown>=3.9",
  "pillow>=11.0.0",
  "python-dotenv>=1.2.1",
  "python-multipart>=0.0.20",
  "pyyaml>=6.0.3",
  "uvicorn>=0.39.0",
]

[dependency-groups]
dev = [
  "httpx>=0.28.1",
  "pytest>=8.4.2",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
